// TakosEncryptInk/types.ts
type Sign = {
  signature: string // ArrayBufferをbase64に変換したもの
  hashedPublicKeyHex: string // 公開鍵をハッシュ化し、16進数文字列に変換したもの
  version: number // 署名のバージョン
}
// ...
interface RoomKey {
  key: string
  sign: Sign
  keyType: "roomKey"
  timestamp: string // 鍵の作成日時
  hashHex: string
  version: number
  masterKeysHashHex: string[]
  roomid: string
}
// ...
interface migrateDataSignKeyPub {
  key: string
  keyType: "migrateDataSignPub"
  version: number
}